[actor #173] refactor
